Migrated update connection test to jest
